Copy provided hints to safely add required DEFAULT_NAMESPACE_HINT_KEY and VALIDATION_HINT [GEOT-5514] (#1338)
* Copy provided hints to safely add required DEFAULT_NAMESPACE_HINT_KEY and VALIDATION_HINT [GEOT-5514]

For details see: https://osgeo-org.atlassian.net/browse/GEOT-5514

The change is needed to respect the ENTITY_RESOLVER hint provided when this code is used by GeoWebCache.

* Clean up warnings on hints use

* respect validation hint if passed in

Cleaned up test story to verify that validation is respected and that entity resolver works on wider range of sample getcapabilities documents.

* Explicitly mention verion negotiation failing when unable to connect

This is more specific than cannot connect, which could imply that the URL was unable to connect.

# Conflicts:
#	modules/unsupported/wfs/src/test/java/org/geotools/data/wfs/v1_0_0/ParseCapabilitesTest.java